<Feature> Map work, Consumables
Added animal icon to map.
Added open inventory hint over discovered resources.
Added region drop rate for resources.
Improved inventory when exploring. Player can now open inventory and eat
food.
Updated compass to highlight resources on compass pulse contact.
Hidden resources no longer glow.
Fixed issue where attributemodifier would not apply from food.
Added inventory limit for player.
Added lots of text events.
Removed thirst and hunger modifiers.
Updated loot frop from enemies to be enemy specific.
Fixed issue where enemies would get stuck in suspicious behaviour.
Fixed issue getting out of bounds grid cell in pathing grid.
Added Inventory and Compass keys.
Removed hold-r to open inventory. Ui no longer fades in and out.
Fixed issue where time between regions was not calculating correctly.
Fixed issue in region names by stripping whitespace from the xml.
Fixed issue where scenery was not updated when game was paused.
Changed inventory to remove resources when they have a value of 0.
Inventory is smaller and simpler now.
Optimised inputhandler.
Fixed issue where it was possible to consume a null resource when the
player has no consumables form the consumable menu.
Updated balance files.
Fixed bug in xslx-xml python file.

+++ Night Unity Files/Assets/Scripts/SamsHelper/ReactiveUI/Elements/EnhancedText.cs
@@ -22,7 +22,7 @@ namespace SamsHelper.ReactiveUI.Elements
         }
 
         private TextMeshProUGUI _text;
-        public int CustomFontSize;
+        public float CustomFontSize;
 
         public FontSizes FontSize;
